6.-Commit
Here I have created all the necessary classes to insert data into the Genre, Platform, Products, Products_has_Genre and User_Address tables of the database. I have also created methods which encrypt the user's password and the methods which check that there are no repeated values when trying to insert the data.